Handle swatch preview
