+++ src/attributes/components/AttributeSwatchField/useFileProcessing.test.tsx
@@ -2,7 +2,7 @@ import { useFileUploadMutation } from "@dashboard/graphql";
 import { act, renderHook } from "@testing-library/react-hooks";
 import React from "react";
 
-import { useFileProcessing } from "./AttributeSwatchField";
+import { useFileProcessing } from "./useFileProcessing";
 
 jest.mock("@dashboard/graphql", () => ({
   useFileUploadMutation: jest.fn(),
